Update Invoke-SlackBot.ps1
Required modules need to be specified in the manifest, not imported later upon function execution.

Also, as currently written your file paths were relative to whatever directory you happened to be in at the time of function execution. I've changed the references to default to the root of module folder. I've also parameterized the path to windows.xml