Update Invoke-SlackBot.ps1
Line 54  -  Fixes a bug I found (can't toLower() on a null) which occurs when you use the newly added Start a Thread feature to reply to a comment rather than a channel. This was crashing the bot.

Line 55-58  -  Trims out consecutive spaces, and null array entries created by trailing spaces to give a cleaner word-array for better follow-on behavior.